fix hbao bug

+++ Core/Scenes/Camera.cpp
@@ -103,9 +103,9 @@ const uint32_t Camera::GetPixelWidth() const
 {
     return Devices::Get()->GetWindow()->GetSize().y;
 }
-const glm::uvec4 Camera::GetPixelSize() const
+const glm::vec4 Camera::GetPixelSize() const
 {
-    return glm::uvec4(GetPixelWidth(), GetPixelHeight(), 1.0f / GetPixelWidth(), 1.0f / GetPixelHeight());
+    return glm::vec4(GetPixelWidth(), GetPixelHeight(), 1.0f / GetPixelWidth(), 1.0f / GetPixelHeight());
 }
 
 const uint32_t Camera::GetStereoPixelHeight() const
@@ -116,8 +116,8 @@ const uint32_t Camera::GetStereoPixelWidth() const
 {
     return Devices::Get()->GetWindow()->GetStereoSize().x;
 }
-const glm::uvec4 Camera::GetStereoPixelSize() const
+const glm::vec4 Camera::GetStereoPixelSize() const
 {
-    return glm::uvec4(GetStereoPixelWidth(), GetStereoPixelHeight(), 1.0f / GetStereoPixelWidth(), 1.0f / GetStereoPixelHeight());
+    return glm::vec4(GetStereoPixelWidth(), GetStereoPixelHeight(), 1.0f / GetStereoPixelWidth(), 1.0f / GetStereoPixelHeight());
 }
 }   // namespace MapleLeaf
